+++ module_board/__init__.py
@@ -1,0 +1,1 @@
+from .board import Board
