+++ module_output/__init__.py
@@ -1,0 +1,1 @@
+from .pdf_service import PdfPrinter
