+++ module_problem_generator/__init__.py
@@ -1,0 +1,5 @@
+from .difficulties import *
+from .module_problem_base_class import *
+from .module_brute_force_backtracking import BruteForceBacktracking
+from .module_dynamic_backtracking import DynamicBacktracking
+from .module_recursive_backtracking import RecursiveBacktracking
