+++ module_problem_generator/module_dynamic_backtracking.py
@@ -2,9 +2,9 @@ import math
 import random
 import copy
 
-import Board.module_board
-import ProblemGenerator.difficulties as difficulties
-from ProblemGenerator.module_problem_base_class import ProblemFinder
+from module_board import Board
+import module_problem_generator.difficulties as difficulties
+from module_problem_generator.module_problem_base_class import ProblemFinder
 
 
 def get_identifier(x: int, y: int):
