+++ module_problem_generator/module_randome_problem.py
@@ -2,7 +2,7 @@
 from random import seed
 
 from Board.module_board import Board
-from ProblemGenerator.module_problem_base_class import ProblemFinder
+from module_problem_generator.module_problem_base_class import ProblemFinder
 
 
 class RandomProblem(ProblemFinder):
